refactor: Moved checklist files
Moved checklist files from view to components
